% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{A}
\alias{A}
\title{A graphs
List all possible A graphs based on given SIX nodes}
\usage{
A(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of A graphs
}
\description{
A graphs
List all possible A graphs based on given SIX nodes
}
\examples{
A(c(1:6))
}
