% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antenna.R
\name{antenna}
\alias{antenna}
\title{antenna graphs
List all possible antenna graphs based on given SIX nodes}
\usage{
antenna(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of antenna graphs
}
\description{
antenna graphs
List all possible antenna graphs based on given SIX nodes
}
\examples{
antenna(c(1:6))
}
