% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bull.R
\name{bull}
\alias{bull}
\title{bull graphs
List all possible bull graphs based on given FIVE nodes}
\usage{
bull(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of bull graphs
}
\description{
bull graphs
List all possible bull graphs based on given FIVE nodes
}
\examples{
bull(c(1:5))
}
