% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly.R
\name{butterfly}
\alias{butterfly}
\title{butterfly graphs
List all possible butterfly subgraphs based on given FIVE nodes. Note that butterfly is the complement graph of co-butterfly based on give five nodes.}
\usage{
butterfly(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of butterfly graphs
}
\description{
butterfly graphs
List all possible butterfly subgraphs based on given FIVE nodes. Note that butterfly is the complement graph of co-butterfly based on give five nodes.
}
\examples{
butterfly(c(1:5))
}
