% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4.R
\name{c4}
\alias{c4}
\title{c4 graphs
List all possible c4 graphs based on given FOUR nodes}
\usage{
c4(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of c4 graphs
}
\description{
c4 graphs
List all possible c4 graphs based on given FOUR nodes
}
\examples{
c4(c(1:4))
}
