% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c5.R
\name{c5}
\alias{c5}
\title{c5 graphs
List all possible c5 graphs based on given FIVE nodes}
\usage{
c5(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of c5 graphs
}
\description{
c5 graphs
List all possible c5 graphs based on given FIVE nodes
}
\examples{
c5(c(1:5))
}
