% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c6.R
\name{c6}
\alias{c6}
\title{c6 graphs
List all possible R graphs based on given SIX nodes}
\usage{
c6(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of c6 graphs
}
\description{
c6 graphs
List all possible R graphs based on given SIX nodes
}
\examples{
c6(c(1:6))
}
