% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chair.R
\name{chair}
\alias{chair}
\title{chair graphs
List all possible chair graphs based on given FIVE nodes}
\usage{
chair(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of chair graphs
}
\description{
chair graphs
List all possible chair graphs based on given FIVE nodes
}
\examples{
chair(c(1:5))
}
