% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUnique.R
\name{checkUnique}
\alias{checkUnique}
\title{Check duplication
Among the isomorphisms of a certain subgraphs, are there any duplications?}
\usage{
checkUnique(mat)
}
\arguments{
\item{mat}{The matrix with colnames representing as edges, e.g 1-2, 2-3, and its entries are 1, 0 to represents there are edges or not.}
}
\value{
Returns TRUE if there is no duplications among these isomorphisms of the certain subgraphs. Retures the index of duplication if there are duplications.
}
\description{
Check duplication
Among the isomorphisms of a certain subgraphs, are there any duplications?
}
\examples{
checkUnique(fillprob(1:6, c6(1:6)))
}
