% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claw.R
\name{claw}
\alias{claw}
\title{claw graphs
List all possible claw subgraphs based on given nodes}
\usage{
claw(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of claw graphs
}
\description{
claw graphs
List all possible claw subgraphs based on given nodes
}
\examples{
claw(c(1:6))
}
