% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claw_k1.bar.R
\name{claw_k1.bar}
\alias{claw_k1.bar}
\title{claw_k1.bar graphs
List all possible claw_k1.bar graphs based on given nodes nodes. Claw_k1.bar is the complement graph of claw_k1 base on given five nodes}
\usage{
claw_k1.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of claw claw_k1.bar graphs
}
\description{
claw_k1.bar graphs
List all possible claw_k1.bar graphs based on given nodes nodes. Claw_k1.bar is the complement graph of claw_k1 base on given five nodes
}
\examples{
claw_k1.bar(c(1:5))
}
