% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co-butterfly.R
\name{co.butterfly}
\alias{co.butterfly}
\title{co-butterfly graphs
List all possible co-butterfly subgraphs based on given FIVE nodes. Note that co-butterfly is the complement graph of butterfly based on give five nodes.}
\usage{
co.butterfly(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of co-butterfly graphs
}
\description{
co-butterfly graphs
List all possible co-butterfly subgraphs based on given FIVE nodes. Note that co-butterfly is the complement graph of butterfly based on give five nodes.
}
\examples{
co_butterfly(c(1:5))
}
