% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co-dart.R
\name{co.dart}
\alias{co.dart}
\title{co-dart graphs
List all possible co-dart graphs based on FIVE given nodes.}
\usage{
co.dart(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of co-dart graphs
}
\description{
co-dart graphs
List all possible co-dart graphs based on FIVE given nodes.
}
\examples{
co.dart(1:5)
}
