% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co-fork.R
\name{co.fork}
\alias{co.fork}
\title{co-fork graphs
List all possible co-fork graphs based on given Five nodes. Note that co-fork is the complement graphs of fork based on given five nodes.}
\usage{
co.fork(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of co-fork graphs
}
\description{
co-fork graphs
List all possible co-fork graphs based on given Five nodes. Note that co-fork is the complement graphs of fork based on given five nodes.
}
\examples{
co-fork(c(1:5))
}
