% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricket.R
\name{cricket}
\alias{cricket}
\title{cricket graphs
List all possible bull graphs based on given FIVE nodes}
\usage{
cricket(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of cricket graphs
}
\description{
cricket graphs
List all possible bull graphs based on given FIVE nodes
}
\examples{
cricket(c(1:5))

}
