% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{cross graphs
List all possible cross graphs based on given FIVE nodes}
\usage{
cross(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of cross graphs
}
\description{
cross graphs
List all possible cross graphs based on given FIVE nodes
}
\examples{
cross(c(1:6))
}
