% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{dart}
\alias{dart}
\title{dart graphs
List all possible dart graphs based on given Five nodes.}
\usage{
dart(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of dart graphs
}
\description{
dart graphs
List all possible dart graphs based on given Five nodes.
}
\examples{
dart(1:5)
}
