% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deToIn.R
\name{deToIn}
\alias{deToIn}
\title{Increasing order
function change the decreasing pair to increasing, eg. 2-1 to 1-2}
\usage{
deToIn(a, b)
}
\arguments{
\item{a}{One node}

\item{b}{Another node}
}
\value{
A string represending a edge, eg. 1-2
}
\description{
Increasing order
function change the decreasing pair to increasing, eg. 2-1 to 1-2
}
\examples{
deToIn(2,1)
}
