% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond.R
\name{diamond}
\alias{diamond}
\title{Diamond graphs
List all possible fiveFlower graphs based on given Four nodes}
\usage{
diamond(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of Diamond graphs
}
\description{
Diamond graphs
List all possible fiveFlower graphs based on given Four nodes
}
\examples{
diamond(c(1:4))
}
