% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino.R
\name{domino}
\alias{domino}
\title{domino graphs
List all possible domino graphs based on given SIX nodes}
\usage{
domino(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of domino graphs
}
\description{
domino graphs
List all possible domino graphs based on given SIX nodes
}
\examples{
domino(c(1:6))
}
