% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E}
\alias{E}
\title{E graphs
List all possible E graphs based on given SIX nodes}
\usage{
E(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of E graphs
}
\description{
E graphs
List all possible E graphs based on given SIX nodes
}
\examples{
E(c(1:6))
}
