% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillprob.R
\name{fillprob}
\alias{fillprob}
\title{All subgraphs
List all subgraphs of specific number of nodes}
\usage{
fillprob(NN, x)
}
\arguments{
\item{NN}{Possible nodes.}

\item{x}{The matrix shows edges with repect to each subgraph of the complete graph. All entries of the matrix are strings, for example "1-3", which represent endpoints of the edge in increasing order. If it is not in increassing order, the function will correct the order.}
}
\value{
A matrix filled with 1 and 0. Number 1 represents there is an edge in the subgraph, otherwise 0.
}
\description{
All subgraphs
List all subgraphs of specific number of nodes
}
\examples{
fillprob(c(1:5), rbind(c("1-2", "1-3"), c("2-4", "1-3")))
}
