% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNodes.R
\name{findNode}
\alias{findNode}
\title{Find the Node
List a matrix to show the nodes in the graphs according to the times of being connected}
\usage{
findNode(times, graph)
}
\arguments{
\item{times}{A numeric shows the times of nodes which have been connected}

\item{graph}{A vector or matrix shows the graphs. If the graph is a matrix, the nodes of these graphs should be the same}
}
\description{
Find the Node
List a matrix to show the nodes in the graphs according to the times of being connected
}
\examples{
findNode(1, p3(1:3))
findNode(2, p3(1:3))
}
