% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish}
\alias{fish}
\title{fish graphs
List all possible fish graphs based on given SIX nodes}
\usage{
fish(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of fish graphs
}
\description{
fish graphs
List all possible fish graphs based on given SIX nodes
}
\examples{
fish(c(1:6))
}
