% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-pan.R
\name{five_pan}
\alias{five_pan}
\title{5-pan graphs
List all possible 5-pan graphs based on given SIX nodes}
\usage{
five_pan(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of 5-pan graphs
}
\description{
5-pan graphs
List all possible 5-pan graphs based on given SIX nodes
}
\examples{
five_pan(c(1:6))
}
