% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem.R
\name{gem}
\alias{gem}
\title{gem graphs
List all possible gem graphs based on given FIVE nodes}
\usage{
gem(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of gem graphs
}
\description{
gem graphs
List all possible gem graphs based on given FIVE nodes
}
\examples{
gem(c(1:6))
}
