% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.R
\name{H}
\alias{H}
\title{H graphs
List all possible H graphs based on given SIX nodes}
\usage{
H(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of H graphs
}
\description{
H graphs
List all possible H graphs based on given SIX nodes
}
\examples{
H(c(1:6))
}
