% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house.R
\name{house}
\alias{house}
\title{house graphs
List all possible house subgraphs based on given FIVE nodes. Note that house is the complement graph of p5 based on give five nodes.}
\usage{
house(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of house graphs
}
\description{
house graphs
List all possible house subgraphs based on given FIVE nodes. Note that house is the complement graph of p5 based on give five nodes.
}
\examples{
house(c(1:5))
}
