% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNeighbor.R
\name{is.neighbor}
\alias{is.neighbor}
\title{Neighbor
Whether two nodes are neighbors in a graph}
\usage{
is.neighbor(node1, node2, graph)
}
\arguments{
\item{node1}{One node, numeric}

\item{node2}{Another node, numeric}

\item{graph}{A graph, vector}
}
\value{
TRUE if node1 and node2 are neighbors in the graph
}
\description{
Neighbor
Whether two nodes are neighbors in a graph
}
\examples{
is.neighbor(1, 2, c6(1:6)[1, ])
}
