% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k14.R
\name{k14}
\alias{k14}
\title{k14 graphs
List all possible k14 graphs based on given FIVE nodes}
\usage{
k14(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of k14 graphs
}
\description{
k14 graphs
List all possible k14 graphs based on given FIVE nodes
}
\examples{
k14(c(1:5))
}
