% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2.R
\name{k2}
\alias{k2}
\title{K2 edges
List all possible k2 graphs based on given nodes}
\usage{
k2(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of K2 graphs
}
\description{
K2 edges
List all possible k2 graphs based on given nodes
}
\examples{
k2(c(1, 2, 4, 5))
k2(c(1, 4, 3, 5))
}
