% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k23.R
\name{k23}
\alias{k23}
\title{k23 graphs
List all possible k23 graphs based on FIVE given nodes. k2 U k3 is the complement of k23}
\usage{
k23(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of k23 graphs
}
\description{
k23 graphs
List all possible k23 graphs based on FIVE given nodes. k2 U k3 is the complement of k23
}
\examples{
k23(1:5)
}
