% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2_k3.R
\name{k2_k3}
\alias{k2_k3}
\title{k2 U k3 graphs
List all possible k2 U k3 graphs based on FIVE given nodes. k2 U k3 is the complement of k2,3}
\usage{
k2_k3(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of k2 U k3 graphs
}
\description{
k2 U k3 graphs
List all possible k2 U k3 graphs based on FIVE given nodes. k2 U k3 is the complement of k2,3
}
\examples{
k2_k3(1:5)
}
