% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k3.R
\name{k3}
\alias{k3}
\title{K3 graphs
List all possible k3 graphs based on given nodes}
\usage{
k3(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of k3 graphs
}
\description{
K3 graphs
List all possible k3 graphs based on given nodes
}
\examples{
k3(c(1, 2, 4, 5))
k3(c(1, 4, 3, 5))
}
