% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k3_2k1.bar.R
\name{k3_2k1.bar}
\alias{k3_2k1.bar}
\title{k3_2k1.bar graphs
List all possible k3_2k1.bar graphs based on given nodes nodes. It is the complement graph of k3_2k1.}
\usage{
k3_2k1.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of k3_2k1.bar graphs
}
\description{
k3_2k1.bar graphs
List all possible k3_2k1.bar graphs based on given nodes nodes. It is the complement graph of k3_2k1.
}
\examples{
k3_2k1.bar(c(1:5))
}
