% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k4.R
\name{k4}
\alias{k4}
\title{k4 graphs
List all possible k4 graphs based on given nodes >= 4}
\usage{
k4(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of k4 graphs
}
\description{
k4 graphs
List all possible k4 graphs based on given nodes >= 4
}
\examples{
k4(c(1:6))
k4(1:4)
}
