% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k5_k1.bar.R
\name{k5_k1.bar}
\alias{k5_k1.bar}
\title{fiveFlower graphs
List all possible fiveFlower graphs based on given SIX nodes}
\usage{
k5_k1.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of fiveFlower graphs
}
\description{
fiveFlower graphs
List all possible fiveFlower graphs based on given SIX nodes
}
\examples{
k5_k1.bar(c(1:6))
}
