% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P.bar.R
\name{P.bar}
\alias{P.bar}
\title{P.bar graphs
List all possible P.bar graphs based on given FIVE nodes}
\usage{
P.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of P.bar graphs
}
\description{
P.bar graphs
List all possible P.bar graphs based on given FIVE nodes
}
\examples{
P.bar(c(1:5))
}
