% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P.R
\name{P}
\alias{P}
\title{P graphs
List all possible P graphs based on given FIVE nodes}
\usage{
P(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of P graphs
}
\description{
P graphs
List all possible P graphs based on given FIVE nodes
}
\examples{
P(c(1:5))
}
