% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_p3.bar.R
\name{p2_p3.bar}
\alias{p2_p3.bar}
\title{p2_p3.bar graphs
List all possible p2_p3.bar graphs based on given nodes nodes. It is the complement graph of p2_p3.}
\usage{
p2_p3.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of p2_p3.bar graphs
}
\description{
p2_p3.bar graphs
List all possible p2_p3.bar graphs based on given nodes nodes. It is the complement graph of p2_p3.
}
\examples{
p2_p3.bar(c(1:5))
}
