% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_p3.R
\name{p2_p3}
\alias{p2_p3}
\title{p2_p3 graphs
List all possible p2_p3 graphs based on given nodes nodes.}
\usage{
p2_p3(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of p2_p3 graphs
}
\description{
p2_p3 graphs
List all possible p2_p3 graphs based on given nodes nodes.
}
\examples{
p2_p3(c(1:5))
}
