% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3.R
\name{p3}
\alias{p3}
\title{P3 edges
List all possible p3 graphs based on given nodes}
\usage{
p3(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of p3 graphs
}
\description{
P3 edges
List all possible p3 graphs based on given nodes
}
\examples{
p3(c(1, 2, 4, 5))
p3(c(1, 4, 3, 5))
}
