% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4.R
\name{p4}
\alias{p4}
\title{p4 graphs
List all possible p4 graphs based on given four nodes}
\usage{
p4(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of p4 graphs
}
\description{
p4 graphs
List all possible p4 graphs based on given four nodes
}
\examples{
p4(c(1:4))
}
