% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p5.R
\name{p5}
\alias{p5}
\title{p5 graphs
List all possible p5 subgraphs based on given FIVE nodes}
\usage{
p5(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of p5 graphs
}
\description{
p5 graphs
List all possible p5 subgraphs based on given FIVE nodes
}
\examples{
p5(c(1:5))
}
