% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p6.R
\name{p6}
\alias{p6}
\title{p6 graphs
List all possible p6 subgraphs based on given FIVE nodes}
\usage{
p6(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of p6 graphs
}
\description{
p6 graphs
List all possible p6 subgraphs based on given FIVE nodes
}
\examples{
p6(c(1:6))
}
