% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{paw}
\alias{paw}
\title{paw graphs
List all possible claw subgraphs based on given FOUR nodes}
\usage{
paw(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of paw graphs
}
\description{
paw graphs
List all possible claw subgraphs based on given FOUR nodes
}
\examples{
paw(c(1:4))
}
