% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_mobius.R
\name{prob_mobius}
\alias{prob_mobius}
\title{Mobius parameters
Base on the binary data probability space, all the Mobius parameters, or marginal probability are represented.}
\usage{
prob_mobius(prob_space)
}
\arguments{
\item{prob_space}{The matrix with 0 and 1 entries to list all possible binary data}
}
\value{
Returns a matrix
}
\description{
Mobius parameters
Base on the binary data probability space, all the Mobius parameters, or marginal probability are represented.
}
\examples{

prob_mobius(prob_space.mat)
}
