% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probToEdges.R
\name{probToEdges}
\alias{probToEdges}
\title{From probability to edges}
\usage{
probToEdges(x)
}
\arguments{
\item{x}{The matrix, whose colomn names represents the edges, is written as 1 and 0, where 1 means there is edge.}
}
\value{
A matrix, each rows represent edges of graphs
}
\description{
From probability to edges
}
\examples{
mat = fillprob(N=4, c4(c(1:4)))
probToEdges(mat)
}
