% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{R}
\alias{R}
\title{R graphs
List all possible R graphs based on given SIX nodes}
\usage{
R(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of R graphs
}
\description{
R graphs
List all possible R graphs based on given SIX nodes
}
\examples{
R(c(1:6))
}
