% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.bar.R
\name{s3.bar}
\alias{s3.bar}
\title{s3.bar graphs
List all possible R graphs based on given SIX nodes}
\usage{
s3.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of s3.bar graphs
}
\description{
s3.bar graphs
List all possible R graphs based on given SIX nodes
}
\examples{
s3.bar(c(1:6))
}
