% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeK2.R
\name{threeK2}
\alias{threeK2}
\title{3k2 graphs
List all possible 3k2 graphs based on given six nodes}
\usage{
threeK2(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of 3k2 graphs
}
\description{
3k2 graphs
List all possible 3k2 graphs based on given six nodes
}
\examples{
threek2(c(1:6))
}
