% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin-c5.R
\name{twin.c5}
\alias{twin.c5}
\title{twin-c5 graphs
List all possible twin-c5 graphs based on given SIX nodes}
\usage{
twin.c5(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of twin-c5 graphs
}
\description{
twin-c5 graphs
List all possible twin-c5 graphs based on given SIX nodes
}
\examples{
domtwin.c5ino(c(1:6))
}
