% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin-house.bar.R
\name{twin.house.bar}
\alias{twin.house.bar}
\title{twin-house.bar graphs
List all possible twin-house.bar graphs based on given SIX nodes}
\usage{
twin.house.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of twin-house.bar graphs
}
\description{
twin-house.bar graphs
List all possible twin-house.bar graphs based on given SIX nodes
}
\examples{
twin.house.bar(c(1:6))
}
