% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoK2.R
\name{twoK2}
\alias{twoK2}
\title{2k2 graphs
List all possible 2k2 graphs based on given nodes}
\usage{
twoK2(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of 2k2 graphs
}
\description{
2k2 graphs
List all possible 2k2 graphs based on given nodes
}
\examples{
2k2(c(1:6))
}
