% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2k3.R
\name{twoK3}
\alias{twoK3}
\title{2k3 graphs
List all possible R graphs based on given SIX nodes}
\usage{
twoK3(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of 2k3 graphs
}
\description{
2k3 graphs
List all possible R graphs based on given SIX nodes
}
\examples{
twoK3(c(1:6))
}
