% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2k3+e.R
\name{twoK3_e}
\alias{twoK3_e}
\title{2K3+e graphs
List all possible 2K3+e graphs based on given SIX nodes}
\usage{
twoK3_e(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of 2K3+e graphs
}
\description{
2K3+e graphs
List all possible 2K3+e graphs based on given SIX nodes
}
\examples{
twoK3_e(c(1:6))
}
