% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twpP3.R
\name{twoP3}
\alias{twoP3}
\title{2p3 graphs
List all possible 2p3 graphs based on given SIX nodes}
\usage{
twoP3(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of 2p3 graphs
}
\description{
2p3 graphs
List all possible 2p3 graphs based on given SIX nodes
}
\examples{
twoP3(c(1:6))
}
