% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w4_k1.bar.R
\name{w4_k1.bar}
\alias{w4_k1.bar}
\title{w4_k1.bar graphs
List all possible w4_k1.bar graphs based on given SIX nodes}
\usage{
w4_k1.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of w4_k1.bar graphs
}
\description{
w4_k1.bar graphs
List all possible w4_k1.bar graphs based on given SIX nodes
}
\examples{
w4_k1.bar(c(1:6))
}
