% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x163.R
\name{x163}
\alias{x163}
\title{x163 graphs
List all possible x163 graphs based on given SIX nodes}
\usage{
x163(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x163 graphs
}
\description{
x163 graphs
List all possible x163 graphs based on given SIX nodes
}
\examples{
x163(c(1:6))
}
