% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x166.R
\name{x166}
\alias{x166}
\title{x166 graphs
List all possible x166 graphs based on given SIX nodes}
\usage{
x166(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x166 graphs
}
\description{
x166 graphs
List all possible x166 graphs based on given SIX nodes
}
\examples{
x166(c(1:6))
}
