% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x167.R
\name{x167}
\alias{x167}
\title{x167 graphs
List all possible x167 graphs based on given SIX nodes}
\usage{
x167(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x167 graphs
}
\description{
x167 graphs
List all possible x167 graphs based on given SIX nodes
}
\examples{
x167(c(1:6))
}
