% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x168.R
\name{x168}
\alias{x168}
\title{x168 graphs
List all possible x168 graphs based on given SIX nodes}
\usage{
x168(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x168 graphs
}
\description{
x168 graphs
List all possible x168 graphs based on given SIX nodes
}
\examples{
x168(c(1:6))
}
