% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x169.R
\name{x169}
\alias{x169}
\title{x169 graphs
List all possible x169 graphs based on given SIX nodes}
\usage{
x169(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x169 graphs
}
\description{
x169 graphs
List all possible x169 graphs based on given SIX nodes
}
\examples{
x169(c(1:6))
}
