% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x170.R
\name{x170}
\alias{x170}
\title{x170 graphs
List all possible x170 graphs based on given SIX nodes}
\usage{
x170(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x170 graphs
}
\description{
x170 graphs
List all possible x170 graphs based on given SIX nodes
}
\examples{
x170(c(1:6))
}
