% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x171.R
\name{x171}
\alias{x171}
\title{x171 graphs
List all possible x171 graphs based on given SIX nodes}
\usage{
x171(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x171 graphs
}
\description{
x171 graphs
List all possible x171 graphs based on given SIX nodes
}
\examples{
x171(c(1:6))
}
