% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x172.R
\name{x172}
\alias{x172}
\title{x172 graphs
List all possible x172 graphs based on given SIX nodes}
\usage{
x172(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x172 graphs
}
\description{
x172 graphs
List all possible x172 graphs based on given SIX nodes
}
\examples{
x172(c(1:6))
}
