% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x18.R
\name{x18}
\alias{x18}
\title{x18 graphs
List all possible x18 graphs based on given SIX nodes}
\usage{
x18(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x18 graphs
}
\description{
x18 graphs
List all possible x18 graphs based on given SIX nodes
}
\examples{
x18(c(1:6))
}
