% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x37.R
\name{x37}
\alias{x37}
\title{x37 graphs
List all possible x37 graphs based on given SIX nodes}
\usage{
x37(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x37 graphs
}
\description{
x37 graphs
List all possible x37 graphs based on given SIX nodes
}
\examples{
x37(c(1:6))
}
