% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x45.R
\name{x45}
\alias{x45}
\title{x45 graphs
List all possible x45 graphs based on given SIX nodes}
\usage{
x45(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x45 graphs
}
\description{
x45 graphs
List all possible x45 graphs based on given SIX nodes
}
\examples{
x45(c(1:6))
}
