% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x5.bar.R
\name{x5.bar}
\alias{x5.bar}
\title{x5.bar graphs
List all possible x5.bar graphs based on given SIX nodes}
\usage{
x5.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x5.bar graphs
}
\description{
x5.bar graphs
List all possible x5.bar graphs based on given SIX nodes
}
\examples{
x5.bar(c(1:6))
}
