% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x58.bar.R
\name{x58.bar}
\alias{x58.bar}
\title{x58.bar graphs
List all possible x58.bar graphs based on given SIX nodes}
\usage{
x58.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x58.bar graphs
}
\description{
x58.bar graphs
List all possible x58.bar graphs based on given SIX nodes
}
\examples{
x58.bar(c(1:6))
}
