% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x84.R
\name{x84}
\alias{x84}
\title{x84 graphs
List all possible x84 graphs based on given SIX nodes}
\usage{
x84(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x84 graphs
}
\description{
x84 graphs
List all possible x84 graphs based on given SIX nodes
}
\examples{
x84(c(1:6))
}
