% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x95.R
\name{x95}
\alias{x95}
\title{x95 graphs
List all possible x95 graphs based on given SIX nodes}
\usage{
x95(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x95 graphs
}
\description{
x95 graphs
List all possible x95 graphs based on given SIX nodes
}
\examples{
x95(c(1:6))
}
