% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x96.R
\name{x96}
\alias{x96}
\title{x96 graphs
List all possible x96 graphs based on given SIX nodes}
\usage{
x96(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x96 graphs
}
\description{
x96 graphs
List all possible x96 graphs based on given SIX nodes
}
\examples{
x96(c(1:6))
}
