% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x98.bar.R
\name{x98.bar}
\alias{x98.bar}
\title{x98.bar graphs
List all possible R graphs based on given SIX nodes}
\usage{
x98.bar(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of x98.bar graphs
}
\description{
x98.bar graphs
List all possible R graphs based on given SIX nodes
}
\examples{
x98.bar(c(1:6))
}
