% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y1.R
\name{y1}
\alias{y1}
\title{Y1 graphs
List all possible Y1 graphs based on given SIX nodes}
\usage{
y1(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of Y1 graphs
}
\description{
Y1 graphs
List all possible Y1 graphs based on given SIX nodes
}
\examples{
y1(c(1:6))
}
