% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2.R
\name{y2}
\alias{y2}
\title{Y2 graphs
List all possible Y2 graphs based on given SIX nodes}
\usage{
y2(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of Y2 graphs
}
\description{
Y2 graphs
List all possible Y2 graphs based on given SIX nodes
}
\examples{
y2(c(1:6))
}
