% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y3.R
\name{y3}
\alias{y3}
\title{y3 graphs
List all possible y3 graphs based on given SIX nodes}
\usage{
y3(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of y3 graphs
}
\description{
y3 graphs
List all possible y3 graphs based on given SIX nodes
}
\examples{
y3(c(1:6))
}
