% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y4.R
\name{y4}
\alias{y4}
\title{y4 graphs
List all possible twin-house.bar graphs based on given SIX nodes}
\usage{
y4(x)
}
\arguments{
\item{x}{The vector representing nodes}
}
\value{
A matrix listing edges of y4 graphs
}
\description{
y4 graphs
List all possible twin-house.bar graphs based on given SIX nodes
}
\examples{
y4(c(1:6))
}
